+++ Sources/VComponents/Components/Buttons/Primary Button/VPrimaryButton.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import VCore
 
 // MARK: - V Primary Button
 /// Large colored button component that performs action when triggered.
@@ -86,7 +87,7 @@ public struct VPrimaryButton<Label>: View where Label: View {
     
     // MARK: Body
     public var body: some View {
-        VBaseButton(gesture: gestureHandler, label: {
+        SwiftUIBaseButton(gesture: gestureHandler, label: {
             buttonLabel
                 .frame(height: uiModel.layout.height)
                 .background(background)
@@ -162,7 +163,7 @@ public struct VPrimaryButton<Label>: View where Label: View {
     }
     
     // MARK: Actions
-    private func gestureHandler(gestureState: VBaseButtonGestureState) {
+    private func gestureHandler(gestureState: BaseButtonGestureState) {
         isPressed = gestureState.isPressed
         if gestureState.isClicked { action() }
     }