+++ Sources/VComponents/Components/Buttons/Secondary Button/VSecondaryButton.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import VCore
 
 // MARK: - V Secondary Button
 /// Small colored button component that performs action when triggered.
@@ -76,7 +77,7 @@ public struct VSecondaryButton<Label>: View where Label: View {
 
     // MARK: Body
     public var body: some View {
-        VBaseButton(gesture: gestureHandler, label: {
+        SwiftUIBaseButton(gesture: gestureHandler, label: {
             buttonLabel
                 .frame(height: uiModel.layout.height)
                 .background(background)
@@ -133,7 +134,7 @@ public struct VSecondaryButton<Label>: View where Label: View {
     }
 
     // MARK: Actions
-    private func gestureHandler(gestureState: VBaseButtonGestureState) {
+    private func gestureHandler(gestureState: BaseButtonGestureState) {
         isPressed = gestureState.isPressed
         if gestureState.isClicked { action() }
     }