+++ Sources/VComponents/Components/State Pickers/Check Box/VCheckBox.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import VCore
 
 // MARK: - V Check Box
 /// State picker component that toggles between off, on, indeterminate, or disabled states, and displays label.
@@ -125,7 +126,7 @@ public struct VCheckBox<Label>: View where Label: View {
                     
                     spacer
 
-                    VBaseButton(gesture: gestureHandler, label: {
+                    SwiftUIBaseButton(gesture: gestureHandler, label: {
                         VText(
                             type: .multiLine(alignment: .leading, lineLimit: uiModel.layout.titleLineLimit),
                             color: uiModel.colors.title.for(internalState),
@@ -142,7 +143,7 @@ public struct VCheckBox<Label>: View where Label: View {
                     
                     spacer
                     
-                    VBaseButton(gesture: gestureHandler, label: {
+                    SwiftUIBaseButton(gesture: gestureHandler, label: {
                         label()
                             .opacity(uiModel.colors.customLabelOpacities.for(internalState))
                     })
@@ -154,7 +155,7 @@ public struct VCheckBox<Label>: View where Label: View {
     }
     
     private var checkBox: some View {
-        VBaseButton(gesture: gestureHandler, label: {
+        SwiftUIBaseButton(gesture: gestureHandler, label: {
             ZStack(content: {
                 RoundedRectangle(cornerRadius: uiModel.layout.cornerRadius)
                     .foregroundColor(uiModel.colors.fill.for(internalState))
@@ -176,7 +177,7 @@ public struct VCheckBox<Label>: View where Label: View {
     }
     
     private var spacer: some View {
-        VBaseButton(gesture: gestureHandler, label: {
+        SwiftUIBaseButton(gesture: gestureHandler, label: {
             Rectangle()
                 .fixedSize(horizontal: false, vertical: true)
                 .frame(width: uiModel.layout.checkBoxLabelSpacing)
@@ -186,7 +187,7 @@ public struct VCheckBox<Label>: View where Label: View {
     }
 
     // MARK: Actions
-    private func gestureHandler(gestureState: VBaseButtonGestureState) {
+    private func gestureHandler(gestureState: BaseButtonGestureState) {
         isPressed = gestureState.isPressed
         if gestureState.isClicked { state.setNextState() }
     }